+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Major improvements:
 
+1. Problem viewer
 
 ### Added: